+++ examples/booking_app/lib/feature_hotel/search/views/search_page.dart
@@ -19,7 +19,7 @@ class SearchPage extends StatefulWidget {
   const SearchPage({Key? key}) : super(key: key);
 
   @override
-  _SearchPageState createState() => _SearchPageState();
+  State<SearchPage> createState() => _SearchPageState();
 }
 
 class _SearchPageState extends State<SearchPage> with TickerProviderStateMixin {