+++ examples/favorites_advanced/rx_bloc_favorites_advanced/lib/base/extensions/async_snapshot.dart
@@ -2,5 +2,5 @@ import 'package:flutter/material.dart';
 
 extension RxBlocAsyncSnapshot<T> on AsyncSnapshot<T> {
   Widget build(Widget Function(T) builder, {Widget? fallback}) =>
-      hasData ? builder(data!) : fallback ?? Container();
+      hasData ? builder(data as T) : fallback ?? Container();
 }