+++ examples/github_search/lib/base/di/app_dependencies.dart
@@ -64,7 +64,8 @@ class AppDependencies {
 
   List<Provider> get _analytics => [
         Provider<FirebaseAnalytics>(
-            create: (context) => FirebaseAnalytics.instance),
+          create: (context) => FirebaseAnalytics.instance,
+        ),
         Provider<FirebaseAnalyticsObserver>(
           create: (context) =>
               FirebaseAnalyticsObserver(analytics: context.read()),